fix: 修复设置交互与协议状态并统一分组详情界面
修复诊断模式确认、跑钞回包等待、电机测试生命周期及校准流程；恢复打印留白预览，优化时间滚轮与密码流程。统一设置分组、详情操作栏、控件留白和入口图标，保留协议确认与导航边界。

已完成主机回归、应用及完整系统构建、升级包资源一致性检查；板上效果与触摸仍需验收。

+++ source/artinchip/test-lvgl/un260/diagnostic/diagnostic.c
@@ -71,8 +71,8 @@ void diagnostic_calibration_end_session(void)
 bool diagnostic_calibration_poll(uint32_t now_ms)
 {
     if (!g_calibration_state.session_active || g_calibration_state.timed_out ||
-        (uint32_t)(now_ms - g_calibration_activity_ms) <
-            DIAGNOSTIC_CALIBRATION_TIMEOUT_MS) {
+        (int32_t)(now_ms - g_calibration_activity_ms) <
+            (int32_t)DIAGNOSTIC_CALIBRATION_TIMEOUT_MS) {
         return false;
     }
